chore(op): rename file_codec_ovm_receipt.rs to receipt_file_codec.rs
