+++ crates/optimism/cli/src/receipt_file_codec.rs
